requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
deprecation==2.1.0
distro==1.9.0
fastapi==0.135.1
gotrue==2.9.1
h11==0.14.0
h2==4.3.0
hpack==4.1.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
packaging==26.0
pillow==12.1.1
postgrest==0.13.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
realtime==1.0.6
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
storage3==0.7.7
StrEnum==0.4.15
supabase==2.3.0
supafunc==0.3.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
websockets==12.0
